requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.1
Django==4.0
django-allauth==0.47.0
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-filter==21.1
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-phonenumber-field==6.0.0
django-rest-auth==0.9.5
django-user-agents==0.4.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
frozenlist==1.3.0
geoip2==4.5.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
maxminddb==2.2.0
multidict==6.0.2
numpy==1.22.1
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
phonenumbers==8.12.39
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sentry-sdk==1.5.3
six==1.16.0
sqlparse==0.4.2
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.8
user-agents==2.2.0
yarl==1.7.2
tinytag
